Use album UPC (Universal Product Code)
The album barcodes can be used to search for that album, a matching result guarantees the validity of the associated artist name. The ISRCSearchService is meant to use the International Standard Recording Code of a track to find associated UPC's, but the ISRC site is not webscrape-friendly, so this does not function at the moment.